chore: group user methods in a dedicated file
